extract interface SessionContextStore
reason: to enable multiple implementations
